+++ CHANGELOG.md
@@ -79,7 +79,7 @@ The changelog for **SwifterSwift**. Also see the [releases](https://github.com/S
 - **NSAttributedString**:
   - Update `applying(attributes: [NSAttributedString.Key : Any], toRangesMatching pattern: String)` to respond `applying(attributes: [NSAttributedString.Key : Any], toRangesMatching pattern: String, options: NSRegularExpression.Options = [])` signature to support regular expression options.  [#791](https://github.com/SwifterSwift/SwifterSwift/pull/791) by [DimaZava](https://github.com/DimaZava)
 - **UIButton**:
- - Refactored `centerTextAndImage(spacing:)`: now can specify the relative position of title and image by setting the new parameter `ImagePosition`. [#807](https://github.com/SwifterSwift/SwifterSwift/pull/807) by [qchenqizhi](https://github.com/qchenqizhi).
+ - Refactored `centerTextAndImage(spacing:)`: now the image position can be above the text, by setting a new parameter `imageAboveText` to `true`. [#807](https://github.com/SwifterSwift/SwifterSwift/pull/807) by [qchenqizhi](https://github.com/qchenqizhi).
 
 ### Deprecated